code/requirement.txt
termcolor>=1.1
tabulate>=0.7.7
tqdm>4.11.1
pyarrow>=0.9.0
pyzmq>=16
tensorflow-gpu == 1.11.0
msgpack
msgpack_numpy
datetime
setproctitle
matplotlib
easydict
pyyaml == 3.13
Cython>=0.19.2
numpy>=1.7.1
scipy>=0.13.2
Pillow
opencv-python

